restructured components and added suggested changes from review
